+++ sdk/network/azure-mgmt-dns/azure/mgmt/dns/v2016_04_01/models/_models_py3.py
@@ -1,5 +1,5 @@
+# pylint: disable=line-too-long,useless-suppression
 # coding=utf-8
-# pylint: disable=too-many-lines
 # --------------------------------------------------------------------------
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License. See License.txt in the project root for license information.
@@ -9,10 +9,9 @@
 
 from typing import Any, Dict, List, Optional, TYPE_CHECKING, Union
 
-from ... import _serialization
+from .._utils import serialization as _serialization
 
 if TYPE_CHECKING:
-    # pylint: disable=unused-import,ungrouped-imports
     from .. import models as _models
 
 
@@ -194,7 +193,7 @@ class PtrRecord(_serialization.Model):
         self.ptrdname = ptrdname
 
 
-class RecordSet(_serialization.Model):  # pylint: disable=too-many-instance-attributes
+class RecordSet(_serialization.Model):
     """Describes a DNS record set (a collection of DNS records with the same name and type).
 
     Variables are only populated by the server, and will be ignored when sending a request.
@@ -315,7 +314,7 @@ class RecordSet(_serialization.Model):  # pylint: disable=too-many-instance-attr
         self.etag = etag
         self.metadata = metadata
         self.ttl = ttl
-        self.fqdn = None
+        self.fqdn: Optional[str] = None
         self.a_records = a_records
         self.aaaa_records = aaaa_records
         self.mx_records = mx_records
@@ -381,7 +380,7 @@ class Resource(_serialization.Model):
     Variables are only populated by the server, and will be ignored when sending a request.
 
     :ivar id: Fully qualified resource ID for the resource. Ex -
-     /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}.  # pylint: disable=line-too-long
+     /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}.
     :vartype id: str
     :ivar name: The name of the resource.
     :vartype name: str
@@ -405,9 +404,9 @@ class Resource(_serialization.Model):
     def __init__(self, **kwargs: Any) -> None:
         """ """
         super().__init__(**kwargs)
-        self.id = None
-        self.name = None
-        self.type = None
+        self.id: Optional[str] = None
+        self.name: Optional[str] = None
+        self.type: Optional[str] = None
 
 
 class SoaRecord(_serialization.Model):
@@ -554,7 +553,7 @@ class TrackedResource(Resource):
     All required parameters must be populated in order to send to server.
 
     :ivar id: Fully qualified resource ID for the resource. Ex -
-     /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}.  # pylint: disable=line-too-long
+     /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}.
     :vartype id: str
     :ivar name: The name of the resource.
     :vartype name: str
@@ -614,7 +613,7 @@ class TxtRecord(_serialization.Model):
         self.value = value
 
 
-class Zone(TrackedResource):  # pylint: disable=too-many-instance-attributes
+class Zone(TrackedResource):
     """Describes a DNS zone.
 
     Variables are only populated by the server, and will be ignored when sending a request.
@@ -622,7 +621,7 @@ class Zone(TrackedResource):  # pylint: disable=too-many-instance-attributes
     All required parameters must be populated in order to send to server.
 
     :ivar id: Fully qualified resource ID for the resource. Ex -
-     /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}.  # pylint: disable=line-too-long
+     /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName}.
     :vartype id: str
     :ivar name: The name of the resource.
     :vartype name: str
@@ -707,9 +706,9 @@ class Zone(TrackedResource):  # pylint: disable=too-many-instance-attributes
         super().__init__(tags=tags, location=location, **kwargs)
         self.etag = etag
         self.max_number_of_record_sets = max_number_of_record_sets
-        self.max_number_of_records_per_record_set = None
+        self.max_number_of_records_per_record_set: Optional[int] = None
         self.number_of_record_sets = number_of_record_sets
-        self.name_servers = None
+        self.name_servers: Optional[List[str]] = None
         self.zone_type = zone_type